Rename: utill 함수 파일명 수정 background > bg
