+++ src/utils/helpers/bgColorConverter.js
